requirements.txt
discord.py[voice]==2.3.2
python-dotenv==1.0.1
aiohttp==3.9.3
aiosqlite==0.19.0
wavelink==3.2.1
Pillow==10.2.0
